requirements.txt
PyQt5==5.15.6
exif==1.3.1
google-api-python-client==2.39.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
